Move GitHub auth out of main file
