+++ auth/github.go
@@ -1,0 +1,109 @@
+package auth
+
+import (
+	"context"
+	"golang.org/x/oauth2"
+	"net/http"
+	"log"
+	"encoding/json"
+	"github.com/google/uuid"
+)
+
+func GitHub(addUser func(user, token string), clientID, clientSecret, organisation string) (login, callback http.HandlerFunc) {
+	ctx := context.Background()
+	conf := &oauth2.Config{
+		ClientID:     clientID,
+		ClientSecret: clientSecret,
+		Scopes:       []string{"user", "read:org"},
+		Endpoint: oauth2.Endpoint{
+			AuthURL:  "https://github.com/login/oauth/authorize",
+			TokenURL: "https://github.com/login/oauth/access_token",
+		},
+	}
+
+	login = func(w http.ResponseWriter, r *http.Request) {
+		url := conf.AuthCodeURL("state", oauth2.AccessTypeOnline)
+
+		http.Redirect(w, r, url, http.StatusFound)
+	}
+
+	callback = func(w http.ResponseWriter, r *http.Request) {
+		code := r.FormValue("code")
+
+		tok, err := conf.Exchange(ctx, code)
+		if err != nil {
+			log.Println(err)
+			return
+		}
+
+		client := conf.Client(ctx, tok)
+
+		user, err := getUser(client)
+		if err != nil {
+			log.Println(err)
+			return
+		}
+
+		inOrg, err := isInOrg(client, organisation)
+		if err != nil {
+			log.Println(err)
+			return
+		}
+
+		if inOrg {
+			token := strId()
+			addUser(user, token)
+
+			http.Redirect(w, r, "/?user="+user+"&token="+token, http.StatusFound)
+		} else {
+			http.Redirect(w, r, "/?error=not_in_org", http.StatusFound)
+		}
+	}
+
+	return login, callback
+}
+
+func getUser(client *http.Client) (string, error) {
+	resp, err := client.Get("https://api.github.com/user")
+	if err != nil {
+		return "", err
+	}
+	defer resp.Body.Close()
+
+	var data struct {
+		Login string `json:"login"`
+	}
+	if err = json.NewDecoder(resp.Body).Decode(&data); err != nil {
+		return "", err
+	}
+
+	return data.Login, nil
+}
+
+func isInOrg(client *http.Client, expectedOrg string) (bool, error) {
+	resp, err := client.Get("https://api.github.com/user/orgs")
+	if err != nil {
+		return false, err
+	}
+	defer resp.Body.Close()
+
+	var data []struct {
+		Login string `json:"login"`
+	}
+	if err = json.NewDecoder(resp.Body).Decode(&data); err != nil {
+		return false, err
+	}
+
+	for _, org := range data {
+		if org.Login == expectedOrg {
+			return true, nil
+		}
+	}
+
+	return false, nil
+}
+
+func strId() string {
+	id, _ := uuid.NewRandom()
+	return id.String()
+}
